refactor: 경로에 맞게 파일 위치 조정
