+++ src/pages/point-shop/index.tsx
