+++ src/pages/point-shop/products/[pointProductId]/detail.tsx
